Fix race condition in concurrent hash map
Load testing revealed that more than the max requests per window
were being served. This was because using methods #computeIfAbsent
and #computeIfPresent in tandem on a concurrent hash map introduces
a race condition: the methods don't synchronise with each other.

The fix is to use only a single method. #merge is a good choice
because it allows you to define a value if the key isn't present in
the map (i.e. a user's first request) and a function to update a
value if the key is present (i.e. calculate remaining tokens and
whether limits are exceeded or not).

Multiple calls to merge will be synchronised, thus eliminating
race conditions.

Some tests are also updated as a result. The main logic changes are
that:
1) we no longer need to use an insertion reference to track whether
a bucket is updated or not;
2) we only set the 'exceededLimit' flag on a token bucket if a user
makes a request while having zero tokens. Reaching zero for the
first time is not, in itself, a limit violation.